app/requirements.txt
django==2.2.3
mysqlclient
django-debug-toolbar
faker
gunicorn
pytest
pytest-django
pytest-cov
setuptools
ipdb
djangorestframework
django-rest-auth
django-rest-framework-jwt
Pillow
